Move universal comparisons to Stack, add some map stuff
Moving the comparison/equality to the Stack module allows the universal
operations to be used as the Eq/Ord instance for `Val`, which allows
those to work properly in Maps.

Map handling has been added to the universal eq/ord, to emulate how they
would be compared as unison values, rather than Map equality.

Functions for mediating between Haskell Map and the unison datatype
representation have been added, for use in decompiling and serializing.
